Integrated Google and KeenASR speech recognition. Reduced minimum Android SDK to 26 for ble and audionlp libraries.

+++ audionlp/src/main/java/io/padium/audionlp/AudioTextResult.kt
@@ -1,3 +1,3 @@
 package io.padium.audionlp
 
-data class AudioTextResult(val phrase: String, val score: Int, val probability: Int)
+data class AudioTextResult(val phrase: String, val score: Double, val probability: Double)
